completed the main checks and modified the structure
