+++ pravega-cli/src/main/java/io/pravega/tools/pravegacli/commands/troubleshoot/TroubleshootCheckCommand.java
@@ -9,9 +9,20 @@
  */
 package io.pravega.tools.pravegacli.commands.troubleshoot;
 
+import io.pravega.controller.server.SegmentHelper;
+import io.pravega.controller.server.rpc.auth.AuthHelper;
+import io.pravega.controller.store.stream.ExtendedStreamMetadataStore;
+import io.pravega.controller.store.stream.StreamStoreFactoryExtended;
 import io.pravega.tools.pravegacli.commands.CommandArgs;
+import io.pravega.tools.pravegacli.commands.utils.CLIControllerConfig;
+import lombok.Cleanup;
+import org.apache.curator.framework.CuratorFramework;
+
+import java.util.concurrent.ScheduledExecutorService;
 
 public class TroubleshootCheckCommand extends TroubleshootCommand {
+
+    protected ExtendedStreamMetadataStore store;
 
     /**
      * Creates a new instance of the Command class.
@@ -25,9 +36,42 @@ public class TroubleshootCheckCommand extends TroubleshootCommand {
     @Override
     public void execute() {
         ensureArgCount(2);
+        final String scope = getCommandArgs().getArgs().get(0);
+        final String streamName = getCommandArgs().getArgs().get(1);
 
-        EpochRecordHistoryTimeSeriesCheck ephtChecker = new EpochRecordHistoryTimeSeriesCheck(getCommandArgs());
-        ephtChecker.check();
+        try {
+            @Cleanup
+            CuratorFramework zkClient = createZKClient();
+            ScheduledExecutorService executor = getCommandArgs().getState().getExecutor();
+
+            SegmentHelper segmentHelper = null;
+            if (getCLIControllerConfig().getMetadataBackend().equals(CLIControllerConfig.MetadataBackends.ZOOKEEPER.name())) {
+                store = StreamStoreFactoryExtended.createZKStore(zkClient, executor);
+            } else {
+                segmentHelper = instantiateSegmentHelper(zkClient);
+                AuthHelper authHelper = AuthHelper.getDisabledAuthHelper();
+                store = StreamStoreFactoryExtended.createPravegaTablesStore(segmentHelper, authHelper, zkClient, executor);
+            }
+
+            GeneralCheck generalChecker = new GeneralCheck(getCommandArgs());
+            ScaleCheck scaleChecker = new ScaleCheck(getCommandArgs());
+            TruncateCheck truncateChecker = new TruncateCheck(getCommandArgs());
+            UpdateCheck updateChecker = new UpdateCheck(getCommandArgs());
+            boolean isConsistent;
+
+            output("\n-------GENERAL CHECKUP-------\n\n");
+
+            isConsistent = generalChecker.check(store, executor);
+            if (!isConsistent) {
+                return;
+            }
+
+            int currentEpoch = store.getActiveEpoch(scope, streamName, null,
+                    true, executor).join().getEpoch();
+
+        } catch (Exception e) {
+            System.err.println("Exception accessing metadata store: " + e.getMessage());
+        }
     }
 
     public static CommandDescriptor descriptor() {
