+++ pravega-cli/src/main/java/io/pravega/tools/pravegacli/commands/troubleshoot/TruncateCheck.java
@@ -9,17 +9,11 @@
  */
 package io.pravega.tools.pravegacli.commands.troubleshoot;
 
-import io.pravega.controller.server.SegmentHelper;
-import io.pravega.controller.server.rpc.auth.AuthHelper;
 import io.pravega.controller.store.stream.ExtendedStreamMetadataStore;
 import io.pravega.controller.store.stream.StoreException;
-import io.pravega.controller.store.stream.StreamStoreFactoryExtended;
 import io.pravega.controller.store.stream.VersionedMetadata;
 import io.pravega.controller.store.stream.records.StreamTruncationRecord;
 import io.pravega.tools.pravegacli.commands.CommandArgs;
-import io.pravega.tools.pravegacli.commands.utils.CLIControllerConfig;
-import lombok.Cleanup;
-import org.apache.curator.framework.CuratorFramework;
 
 import java.util.*;
 import java.util.concurrent.ScheduledExecutorService;
@@ -38,83 +32,65 @@ public class TruncateCheck extends TroubleshootCommand {
 
     }
 
-    public boolean check() {
+    public boolean check(ExtendedStreamMetadataStore store, ScheduledExecutorService executor) {
         ensureArgCount(2);
         final String scope = getCommandArgs().getArgs().get(0);
         final String streamName = getCommandArgs().getArgs().get(1);
         StringBuilder responseBuilder = new StringBuilder();
 
+        StreamTruncationRecord truncationRecord;
+
         try {
-            @Cleanup
-            CuratorFramework zkClient = createZKClient();
-            ScheduledExecutorService executor = getCommandArgs().getState().getExecutor();
+            truncationRecord = store.getTruncationRecord(scope, streamName, null, executor)
+                    .thenApply(VersionedMetadata::getObject).join();
 
-            SegmentHelper segmentHelper = null;
-            if (getCLIControllerConfig().getMetadataBackend().equals(CLIControllerConfig.MetadataBackends.ZOOKEEPER.name())) {
-                store = StreamStoreFactoryExtended.createZKStore(zkClient, executor);
-            } else {
-                segmentHelper = instantiateSegmentHelper(zkClient);
-                AuthHelper authHelper = AuthHelper.getDisabledAuthHelper();
-                store = StreamStoreFactoryExtended.createPravegaTablesStore(segmentHelper, authHelper, zkClient, executor);
-            }
+        } catch (StoreException.DataNotFoundException e) {
+            responseBuilder.append("StreamTruncationRecord is corrupted or unavailable").append("\n");
+            output(responseBuilder.toString());
+            return false;
+        }
 
-            StreamTruncationRecord truncationRecord;
-
-            try {
-                truncationRecord = store.getTruncationRecord(scope, streamName, null, executor)
-                        .thenApply(VersionedMetadata::getObject).join();
-
-            } catch (StoreException.DataNotFoundException e) {
-                responseBuilder.append("StreamTruncationRecord is corrupted or unavailable").append("\n");
-                output(responseBuilder.toString());
-                return false;
-            }
-
-            // If the StreamTruncationRecord is EMPTY then there's no need to check further
-            if (truncationRecord.equals(StreamTruncationRecord.EMPTY)) {
-                output("No error involving truncating.");
-                return true;
-            }
-
-            boolean isConsistent = true;
-
-            // Need to check internal consistency
-            // Updating and segments to delete check
-            if (!truncationRecord.isUpdating()) {
-                if (!truncationRecord.getToDelete().isEmpty()) {
-                    responseBuilder.append("Inconsistency in the StreamTruncationRecord in regards to updating and segments to delete").append("\n");
-                    isConsistent = false;
-                }
-            }
-
-            // Correct segments deletion check
-            Long streamCutMaxSegment = Collections.max(truncationRecord.getStreamCut().keySet());
-            Set<Long> allDelete = truncationRecord.getToDelete();
-            allDelete.addAll(truncationRecord.getDeletedSegments());
-
-            List<Long> badSegments = allDelete.stream()
-                    .filter(segment -> segment >= streamCutMaxSegment)
-                    .collect(Collectors.toList());
-
-            if (!badSegments.isEmpty()) {
-                responseBuilder.append("Inconsistency in the StreamTruncationRecord in regards to segments deletion, " +
-                        "segments ahead of stream cut being deleted").append("\n");
-                isConsistent = false;
-            }
-
-            // Based on consistency, return all records or none
-            if (!isConsistent) {
-                responseBuilder.append(outputTruncation(truncationRecord));
-                output(responseBuilder.toString());
-                return false;
-            }
-
-            output("Consistent with respect to truncating");
-            return true;
-
-        } catch (Exception e) {
-            System.err.println("Exception accessing metadata store: " + e.getMessage());
+        // If the StreamTruncationRecord is EMPTY then there's no need to check further
+        if (truncationRecord.equals(StreamTruncationRecord.EMPTY)) {
+            output("No error involving truncating.");
             return true;
         }
+
+        boolean isConsistent = true;
+
+        // Need to check internal consistency
+        // Updating and segments to delete check
+        if (!truncationRecord.isUpdating()) {
+            if (!truncationRecord.getToDelete().isEmpty()) {
+                responseBuilder.append("Inconsistency in the StreamTruncationRecord in regards to updating and segments to delete").append("\n");
+                isConsistent = false;
+            }
+        }
+
+        // Correct segments deletion check
+        Long streamCutMaxSegment = Collections.max(truncationRecord.getStreamCut().keySet());
+        Set<Long> allDelete = truncationRecord.getToDelete();
+        allDelete.addAll(truncationRecord.getDeletedSegments());
+
+        List<Long> badSegments = allDelete.stream()
+                .filter(segment -> segment >= streamCutMaxSegment)
+                .collect(Collectors.toList());
+
+        if (!badSegments.isEmpty()) {
+            responseBuilder.append("Inconsistency in the StreamTruncationRecord in regards to segments deletion, " +
+                    "segments ahead of stream cut being deleted").append("\n");
+            isConsistent = false;
+        }
+
+        // Based on consistency, return all records or none
+        if (!isConsistent) {
+            responseBuilder.append(outputTruncation(truncationRecord));
+            output(responseBuilder.toString());
+            return false;
+        }
+
+        output("Consistent with respect to truncating");
+        return true;
+
     }
 }
